Polish cluster interface.
The interface now being `CausalCluster` so that things are as before,
with `DefaultCausalCluster` the current only implementation.
